Got rid of size_t vs. int warnings

+++ doc/snippets/test_example.cpp
@@ -60,7 +60,7 @@ struct GePolCavityTest {
 BOOST_FIXTURE_TEST_CASE(size, GePolCavityTest)
 {
     int size = 32;
-    int actualSize = cavity.size();
+    size_t actualSize = cavity.size();
     BOOST_REQUIRE_EQUAL(size, actualSize);
 }
 
@@ -83,7 +83,7 @@ BOOST_FIXTURE_TEST_CASE(volume, GePolCavityTest)
     Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
     Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
     double actualVolume = 0;
-    for ( int i = 0; i < cavity.size(); ++i ) {
+    for ( size_t i = 0; i < cavity.size(); ++i ) {
         actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                             i));
     }
