+++ tests/bi_operators/bi_operators_numerical.cpp
@@ -90,8 +90,8 @@ BOOST_FIXTURE_TEST_CASE(vacuum, NumericalIntegratorTest)
     int dim_read = raw_S_ref.shape[0];
     Eigen::MatrixXd S_reference = Eigen::MatrixXd::Zero(dim_read, dim_read);
     S_reference = getFromRawBuffer<double>(dim_read, dim_read, raw_S_ref.data);
-    for (int i = 0; i < cavity.size(); ++i) {
-        for (int j = 0; j < cavity.size(); ++j) {
+    for (size_t i = 0; i < cavity.size(); ++i) {
+        for (size_t j = 0; j < cavity.size(); ++j) {
             BOOST_REQUIRE_CLOSE(S_reference(i, j), S_results(i, j), 1.0e-12);
         }
     }
@@ -103,8 +103,8 @@ BOOST_FIXTURE_TEST_CASE(vacuum, NumericalIntegratorTest)
     dim_read = raw_D_ref.shape[0];
     Eigen::MatrixXd D_reference = Eigen::MatrixXd::Zero(dim_read, dim_read);
     D_reference = getFromRawBuffer<double>(dim_read, dim_read, raw_D_ref.data);
-    for (int i = 0; i < cavity.size(); ++i) {
-        for (int j = 0; j < cavity.size(); ++j) {
+    for (size_t i = 0; i < cavity.size(); ++i) {
+        for (size_t j = 0; j < cavity.size(); ++j) {
             BOOST_REQUIRE_CLOSE(D_reference(i, j), D_results(i, j), 1.0e-12);
         }
     }
@@ -128,8 +128,8 @@ BOOST_FIXTURE_TEST_CASE(uniformdielectric, NumericalIntegratorTest)
     int dim_read = raw_S_ref.shape[0];
     Eigen::MatrixXd S_reference = Eigen::MatrixXd::Zero(dim_read, dim_read);
     S_reference = getFromRawBuffer<double>(dim_read, dim_read, raw_S_ref.data);
-    for (int i = 0; i < cavity.size(); ++i) {
-        for (int j = 0; j < cavity.size(); ++j) {
+    for (size_t i = 0; i < cavity.size(); ++i) {
+        for (size_t j = 0; j < cavity.size(); ++j) {
             BOOST_REQUIRE_CLOSE(S_reference(i, j), S_results(i, j), 1.0e-12);
         }
     }
@@ -141,8 +141,8 @@ BOOST_FIXTURE_TEST_CASE(uniformdielectric, NumericalIntegratorTest)
     dim_read = raw_D_ref.shape[0];
     Eigen::MatrixXd D_reference = Eigen::MatrixXd::Zero(dim_read, dim_read);
     D_reference = getFromRawBuffer<double>(dim_read, dim_read, raw_D_ref.data);
-    for (int i = 0; i < cavity.size(); ++i) {
-        for (int j = 0; j < cavity.size(); ++j) {
+    for (size_t i = 0; i < cavity.size(); ++i) {
+        for (size_t j = 0; j < cavity.size(); ++j) {
             BOOST_REQUIRE_CLOSE(D_reference(i, j), D_results(i, j), 1.0e-12);
         }
     }
@@ -164,8 +164,8 @@ BOOST_FIXTURE_TEST_CASE(ionic, NumericalIntegratorTest)
     int dim_read = raw_S_ref.shape[0];
     Eigen::MatrixXd S_reference = Eigen::MatrixXd::Zero(dim_read, dim_read);
     S_reference = getFromRawBuffer<double>(dim_read, dim_read, raw_S_ref.data);
-    for (int i = 0; i < cavity.size(); ++i) {
-        for (int j = 0; j < cavity.size(); ++j) {
+    for (size_t i = 0; i < cavity.size(); ++i) {
+        for (size_t j = 0; j < cavity.size(); ++j) {
             BOOST_REQUIRE_CLOSE(S_reference(i, j), S_results(i, j), 1.0e-12);
         }
     }
@@ -177,8 +177,8 @@ BOOST_FIXTURE_TEST_CASE(ionic, NumericalIntegratorTest)
     dim_read = raw_D_ref.shape[0];
     Eigen::MatrixXd D_reference = Eigen::MatrixXd::Zero(dim_read, dim_read);
     D_reference = getFromRawBuffer<double>(dim_read, dim_read, raw_D_ref.data);
-    for (int i = 0; i < cavity.size(); ++i) {
-        for (int j = 0; j < cavity.size(); ++j) {
+    for (size_t i = 0; i < cavity.size(); ++i) {
+        for (size_t j = 0; j < cavity.size(); ++j) {
             BOOST_REQUIRE_CLOSE(D_reference(i, j), D_results(i, j), 1.0e-12);
         }
     }
@@ -200,8 +200,8 @@ BOOST_FIXTURE_TEST_CASE(anisotropic, NumericalIntegratorTest)
     int dim_read = raw_S_ref.shape[0];
     Eigen::MatrixXd S_reference = Eigen::MatrixXd::Zero(dim_read, dim_read);
     S_reference = getFromRawBuffer<double>(dim_read, dim_read, raw_S_ref.data);
-    for (int i = 0; i < cavity.size(); ++i) {
-        for (int j = 0; j < cavity.size(); ++j) {
+    for (size_t i = 0; i < cavity.size(); ++i) {
+        for (size_t j = 0; j < cavity.size(); ++j) {
             BOOST_REQUIRE_CLOSE(S_reference(i, j), S_results(i, j), 1.0e-12);
         }
     }
@@ -213,8 +213,8 @@ BOOST_FIXTURE_TEST_CASE(anisotropic, NumericalIntegratorTest)
     dim_read = raw_D_ref.shape[0];
     Eigen::MatrixXd D_reference = Eigen::MatrixXd::Zero(dim_read, dim_read);
     D_reference = getFromRawBuffer<double>(dim_read, dim_read, raw_D_ref.data);
-    for (int i = 0; i < cavity.size(); ++i) {
-        for (int j = 0; j < cavity.size(); ++j) {
+    for (size_t i = 0; i < cavity.size(); ++i) {
+        for (size_t j = 0; j < cavity.size(); ++j) {
             BOOST_REQUIRE_CLOSE(D_reference(i, j), D_results(i, j), 1.0e-12);
         }
     }
