+++ tests/gepol/gepol_C2H4_D2h.cpp
@@ -67,7 +67,7 @@ struct GePolCavityD2hAddTest {
 BOOST_FIXTURE_TEST_CASE(size, GePolCavityD2hAddTest)
 {
     int size = 576;
-    int actualSize = cavity.size();
+    size_t actualSize = cavity.size();
     BOOST_REQUIRE_EQUAL(size, actualSize);
 }
 
@@ -100,7 +100,7 @@ BOOST_FIXTURE_TEST_CASE(volume, GePolCavityD2hAddTest)
     Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
     Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
     double actualVolume = 0;
-    for ( int i = 0; i < cavity.size(); ++i ) {
+    for ( size_t i = 0; i < cavity.size(); ++i ) {
         actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                             i));
     }
