+++ tests/gepol/gepol_H3+_C2v.cpp
@@ -68,7 +68,7 @@ BOOST_FIXTURE_TEST_SUITE(GePolCavityC2vAdd, GePolCavityC2vAddTest)
 BOOST_FIXTURE_TEST_CASE(size, GePolCavityC2vAddTest)
 {
     int size = 312;
-    int actualSize = cavity.size();
+    size_t actualSize = cavity.size();
     BOOST_REQUIRE_EQUAL(size, actualSize);
 }
 
@@ -101,7 +101,7 @@ BOOST_FIXTURE_TEST_CASE(volume, GePolCavityC2vAddTest)
     Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
     Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
     double actualVolume = 0;
-    for ( int i = 0; i < cavity.size(); ++i ) {
+    for ( size_t i = 0; i < cavity.size(); ++i ) {
         actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                             i));
     }
@@ -135,7 +135,7 @@ BOOST_FIXTURE_TEST_SUITE(GePolCavityC2v, GePolCavityC2vTest)
 BOOST_FIXTURE_TEST_CASE(size, GePolCavityC2vTest)
 {
     int size = 288;
-    int actualSize = cavity.size();
+    size_t actualSize = cavity.size();
     BOOST_REQUIRE_EQUAL(size, actualSize);
 }
 
@@ -168,7 +168,7 @@ BOOST_FIXTURE_TEST_CASE(volume, GePolCavityC2vTest)
     Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
     Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
     double actualVolume = 0;
-    for ( int i = 0; i < cavity.size(); ++i ) {
+    for ( size_t i = 0; i < cavity.size(); ++i ) {
         actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                             i));
     }
